Create readmeCollection0.txt
This is a first experimental version of DEM tiles at 2 m resolutions with an extent of 256x256 pixels. The tile cover mainly alpine morphologies (with some exception) in norther Italy, covering the Trentino Province and the Friuli Venezia Giulia Region. The data are derived from public repositories of the provincial and regional institutions.